+++ asphalt/templating/component.py
@@ -13,7 +13,7 @@ logger = logging.getLogger(__name__)
 
 class TemplatingComponent(Component):
     """
-    Publishes one or more :class:`~asphalt.templating.api.TemplateRenderer` lazy resources.
+    Creates one or more :class:`~asphalt.templating.api.TemplateRenderer` resource factories.
 
     If more than one renderer is to be configured, provide a ``renderers`` argument as a dictionary
     where the key is the resource name and the value is a dictionary of keyword arguments to
@@ -26,8 +26,9 @@ class TemplatingComponent(Component):
     based on the provided default arguments, with ``context_attr`` defaulting to the name of the
     chosen backend.
 
-    :param renderers: a dictionary of resource name ⭢ :meth:`configure_renderer` keyword arguments
-    :param default_renderer_args: default values for :meth:`configure_renderer`
+    :param renderers: a dictionary of resource name ⭢ constructor arguments for the chosen backend
+        class
+    :param default_renderer_args: default values for constructor keyword arguments
     """
 
     def __init__(self, renderers: Dict[str, Dict[str, Any]] = None, **default_renderer_args):
@@ -39,30 +40,15 @@ class TemplatingComponent(Component):
         self.renderers = []
         for resource_name, config in renderers.items():
             config = merge_config(default_renderer_args, config or {})
-            config.setdefault('backend', resource_name)
-            config.setdefault('context_attr', resource_name)
-            context_attr, renderer = self.configure_renderer(**config)
+            type_ = config.pop('backend', resource_name)
+            context_attr = config.pop('context_attr', resource_name)
+            renderer = template_renderers.create_object(type_, **config)
             self.renderers.append((resource_name, context_attr, renderer))
-
-    @classmethod
-    def configure_renderer(cls, context_attr: str, backend: str, **backend_args):
-        """
-        Configure a template renderer.
-
-        :param context_attr: context attribute of the renderer (if omitted, the resource name
-            will be used instead)
-        :param backend: entry point name used to look up the backend class (if omitted, the
-            resource name will be used instead)
-        :param backend_args: keyword arguments passed to the constructor of the backend class
-
-        """
-        assert check_argument_types()
-        renderer = template_renderers.create_object(backend, **backend_args)
-        return context_attr, renderer
 
     async def start(self, ctx: Context):
         for resource_name, context_attr, renderer in self.renderers:
             proxymaker = partial(TemplateRendererProxy, renderer=renderer)
-            ctx.publish_lazy_resource(proxymaker, TemplateRenderer, resource_name, context_attr)
+            types = [TemplateRenderer, type(renderer)]
+            ctx.add_resource_factory(proxymaker, types, resource_name, context_attr)
             logger.info('Configured template renderer (%s / ctx.%s; class=%s)', resource_name,
                         context_attr, qualified_name(renderer))