+++ asphalt/templating/renderers/mako.py
@@ -4,7 +4,7 @@ from mako.lookup import TemplateLookup
 from mako.template import Template
 
 from asphalt.templating.api import TemplateRenderer
-from asphalt.templating.util import package_to_directory
+from asphalt.templating.utils import package_to_directory
 
 
 class MakoRenderer(TemplateRenderer):