+++ docs/versionhistory.rst
@@ -3,6 +3,13 @@ Version history
 
 This library adheres to `Semantic Versioning <http://semver.org/>`_.
 
-**1.0.0**
+**2.0.0** (2017-04-11)
+
+- **BACKWARD INCOMPATIBLE** Migrated to Asphalt 3.0
+- **BACKWARD INCOMPATIBLE** Renamed the ``asphalt.templating.util`` module to
+  ``asphalt.templating.utils``
+- Renderer resources are now added to the context using their actual types as well
+
+**1.0.0** (2015-05-21)
 
 - Initial release